+++ src/rematch/mediator/stream_mediator/stream_lbl_mediator.cpp
@@ -32,7 +32,7 @@ std::unique_ptr<mediator::Mapping> StreamLblMediator::next() {
 #endif
   if (next_is_computed_successfully()) {
     auto mapping = construct_user_mapping();
-    mapping->shift(line_offset);
+    mapping->shift(static_cast<int64_t>(line_offset));
     return mapping;
   }
